pyproject.toml
[project]
name = "wplace-auto-painter-pw"
version = "0.1.0"
description = "Some awesome description..."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "anyio>=4.11.0",
  "bot7685-ext==0.3.0",
  "cloudscraper>=1.2.71",
  "httpx>=0.28.1",
  "loguru>=0.7.3",
  "pillow>=12.0.0",
  "playwright==1.55.0",
  "pydantic>=2.12.3",
  "pyqt6>=6.10.0",
  "pyqt6-fluent-widgets>=1.11.1",
  "windows-toasts>=1.3.1 ; sys_platform == 'win32'",
]

[dependency-groups]
dev = ["pyinstaller>=6.16.0"]

[tool.uv]
package = false
python-downloads = "automatic"

[tool.pyright]
pythonVersion = "3.14"
pythonPlatform = "All"
typeCheckingMode = "standard"
ignore = [".venv", "data", "**/build", "**/scripts"]
disableBytesTypePromotions = true
